organize GET video by specific category into videoController, videoRoutes, & videoService from index.js on backend
